extra deciimal for humidity

+++ pmsd001.go
@@ -32,13 +32,12 @@ func getTemp(d i2c.Device) (float64, float64) {
 
 	temp := (int64(response[0])*256 + int64(response[1])) & 0xFFFC
 	tempC := -46.85 + (175.72 * float64(temp) / 65536.0)
-	tempF := tempC*1.8 + 32
+	tempF := tempC * 1.8 + 32
 
-	// fmt.Printf("%.2f °F\n", tempF)
 	return tempF, tempC
 }
 
-func getRelativeHumidity(d i2c.Device) int64 {
+func getRelativeHumidity(d i2c.Device) float64 {
 	err := d.Write([]byte{HUMIDITY})
 	check(err)
 
@@ -47,8 +46,8 @@ func getRelativeHumidity(d i2c.Device) int64 {
 	response := make([]byte, 2)
 	check(d.Read(response))
 
-	data := (int64(response[0])*256 + int64(response[1])) & 0xFFFC
-	humidity := ((125 * data) / 65536) - 6
+	data := (int64(response[0]) * 256 + int64(response[1])) & 0xFFFC
+	humidity := (125 * float64(data) / 65536) - 6
 
 	return humidity
 }
@@ -60,5 +59,5 @@ func main() {
 	tempF, tempC := getTemp(*d)
 	humidity := getRelativeHumidity(*d)
 
-	fmt.Printf("%d\t%.2f\t%.2f\t%d\n", time.Now().Unix(), tempC, tempF, humidity)
+	fmt.Printf("%d\t%.2f\t%.2f\t%.1f\n", time.Now().Unix(), tempC, tempF, humidity)
 }
